upload-image-microservice/requirements.txt
django
djangorestframework
mysqlclient
django-mysql
django-cors-headers
pymysql
cryptography
django-googledrive-storage
pillow